Input validation

+++ shopping/src/main/java/com/spring/controller/AdminController.java
@@ -20,6 +20,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.spring.constants.ResponseCode;
 import com.spring.constants.WebConstants;
 import com.spring.model.PlaceOrder;
 import com.spring.model.Product;
@@ -32,6 +33,7 @@ import com.spring.response.order;
 import com.spring.response.prodResp;
 import com.spring.response.serverResp;
 import com.spring.response.viewOrdResp;
+import com.spring.util.Validator;
 import com.spring.util.jwtUtil;
 
 @CrossOrigin(origins = WebConstants.ALLOWED_URL)
@@ -41,33 +43,41 @@ public class AdminController {
 
 	@Autowired
 	private UserRepository userRepo;
+
 	@Autowired
 	private ProductRepository prodRepo;
+
 	@Autowired
 	private OrderRepository ordRepo;
+
 	@Autowired
 	private CartRepository cartRepo;
+
 	@Autowired
 	private jwtUtil jwtutil;
 
 	@PostMapping("/verify")
 	public ResponseEntity<serverResp> verifyUser(@Valid @RequestBody HashMap<String, String> credential) {
-
-		String email = credential.get(WebConstants.USER_EMAIL);
-		String password = credential.get(WebConstants.USER_PASSWORD);
+		String email = "";
+		String password = "";
+		if (credential.containsKey(WebConstants.USER_EMAIL)) {
+			email = credential.get(WebConstants.USER_EMAIL);
+		}
+		if (credential.containsKey(WebConstants.USER_PASSWORD)) {
+			password = credential.get(WebConstants.USER_PASSWORD);
+		}
 		User loggedUser = userRepo.findByEmailAndPasswordAndUsertype(email, password, WebConstants.USER_ADMIN_ROLE);
 		serverResp resp = new serverResp();
 		if (loggedUser != null) {
 			String jwtToken = jwtutil.createToken(email, password, WebConstants.USER_ADMIN_ROLE);
-			resp.setStatus(WebConstants.SUCCESS_CODE);
-			resp.setMessage(WebConstants.SUCCESS_MESSAGE);
+			resp.setStatus(ResponseCode.SUCCESS_CODE);
+			resp.setMessage(ResponseCode.SUCCESS_MESSAGE);
 			resp.setAUTH_TOKEN(jwtToken);
-			return new ResponseEntity<serverResp>(resp, HttpStatus.OK);
 		} else {
-			resp.setStatus(WebConstants.FAILURE_CODE);
-			resp.setMessage(WebConstants.FAILURE_MESSAGE);
-			return new ResponseEntity<serverResp>(resp, HttpStatus.NOT_ACCEPTABLE);
+			resp.setStatus(ResponseCode.FAILURE_CODE);
+			resp.setMessage(ResponseCode.FAILURE_MESSAGE);
 		}
+		return new ResponseEntity<serverResp>(resp, HttpStatus.OK);
 	}
 
 	@PostMapping("/addProduct")
@@ -78,7 +88,11 @@ public class AdminController {
 			@RequestParam(name = WebConstants.PROD_NAME) String productname,
 			@RequestParam(name = WebConstants.PROD_QUANITY) String quantity) throws IOException {
 		prodResp resp = new prodResp();
-		if (jwtutil.checkToken(AUTH_TOKEN) != null) {
+		if (Validator.isStringEmpty(productname) || Validator.isStringEmpty(description)
+				|| Validator.isStringEmpty(price) || Validator.isStringEmpty(quantity)) {
+			resp.setStatus(ResponseCode.BAD_REQUEST_CODE);
+			resp.setMessage(ResponseCode.BAD_REQUEST_MESSAGE);
+		} else if (!Validator.isStringEmpty(AUTH_TOKEN) && jwtutil.checkToken(AUTH_TOKEN) != null) {
 			try {
 				Product prod = new Product();
 				prod.setDescription(description);
@@ -87,46 +101,43 @@ public class AdminController {
 				prod.setQuantity(Integer.parseInt(quantity));
 				prod.setProductimage(prodImage.getBytes());
 				prodRepo.save(prod);
-				resp.setStatus(WebConstants.ADD_SUCCESS_CODE);
-				resp.setMessage(WebConstants.ADD_SUCCESS_MESSAGE);
+
+				resp.setStatus(ResponseCode.SUCCESS_CODE);
+				resp.setMessage(ResponseCode.ADD_SUCCESS_MESSAGE);
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
 				resp.setOblist(prodRepo.findAll());
-				return new ResponseEntity<prodResp>(resp, HttpStatus.CREATED);
 			} catch (Exception e) {
-				resp.setStatus(WebConstants.ADD_FAILURE_CODE);
+				resp.setStatus(ResponseCode.FAILURE_CODE);
 				resp.setMessage(e.getMessage());
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 			}
 		} else {
-			resp.setStatus(WebConstants.FAILURE_CODE);
-			resp.setMessage(WebConstants.FAILURE_MESSAGE);
-			return new ResponseEntity<prodResp>(resp, HttpStatus.NOT_ACCEPTABLE);
+			resp.setStatus(ResponseCode.BAD_REQUEST_CODE);
+			resp.setMessage(ResponseCode.BAD_REQUEST_MESSAGE);
 		}
+		return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 	}
 
 	@PostMapping("/getProducts")
 	public ResponseEntity<prodResp> getProducts(@RequestHeader(name = WebConstants.USER_AUTH_TOKEN) String AUTH_TOKEN)
 			throws IOException {
 		prodResp resp = new prodResp();
-		if (jwtutil.checkToken(AUTH_TOKEN) != null) {
+		if (!Validator.isStringEmpty(AUTH_TOKEN) && jwtutil.checkToken(AUTH_TOKEN) != null) {
 			try {
-				resp.setStatus(WebConstants.SUCCESS_CODE);
-				resp.setMessage(WebConstants.LIST_SUCCESS_MESSAGE);
+				resp.setStatus(ResponseCode.SUCCESS_CODE);
+				resp.setMessage(ResponseCode.LIST_SUCCESS_MESSAGE);
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
 				resp.setOblist(prodRepo.findAll());
-				return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 			} catch (Exception e) {
-				resp.setStatus(WebConstants.LIST_FAILURE_CODE);
+				resp.setStatus(ResponseCode.FAILURE_CODE);
 				resp.setMessage(e.getMessage());
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 			}
 		} else {
-			resp.setStatus(WebConstants.FAILURE_CODE);
-			resp.setMessage(WebConstants.FAILURE_MESSAGE);
-			return new ResponseEntity<prodResp>(resp, HttpStatus.NOT_ACCEPTABLE);
+			resp.setStatus(ResponseCode.FAILURE_CODE);
+			resp.setMessage(ResponseCode.FAILURE_MESSAGE);
 		}
+		return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 	}
 
 	@PostMapping("/updateProducts")
@@ -139,7 +150,11 @@ public class AdminController {
 			@RequestParam(name = WebConstants.PROD_QUANITY) String quantity,
 			@RequestParam(name = WebConstants.PROD_ID) String productid) throws IOException {
 		serverResp resp = new serverResp();
-		if (jwtutil.checkToken(AUTH_TOKEN) != null) {
+		if (Validator.isStringEmpty(productname) || Validator.isStringEmpty(description)
+				|| Validator.isStringEmpty(price) || Validator.isStringEmpty(quantity)) {
+			resp.setStatus(ResponseCode.BAD_REQUEST_CODE);
+			resp.setMessage(ResponseCode.BAD_REQUEST_MESSAGE);
+		} else if (!Validator.isStringEmpty(AUTH_TOKEN) && jwtutil.checkToken(AUTH_TOKEN) != null) {
 			try {
 				Product prodOrg;
 				Product prod;
@@ -152,47 +167,45 @@ public class AdminController {
 							Integer.parseInt(quantity), prodOrg.getProductimage());
 				}
 				prodRepo.save(prod);
-				resp.setStatus(WebConstants.SUCCESS_CODE);
-				resp.setMessage(WebConstants.UPD_SUCCESS_MESSAGE);
+				resp.setStatus(ResponseCode.SUCCESS_CODE);
+				resp.setMessage(ResponseCode.UPD_SUCCESS_MESSAGE);
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<serverResp>(resp, HttpStatus.ACCEPTED);
 			} catch (Exception e) {
-				resp.setStatus(WebConstants.UPD_FAILURE_CODE);
+				resp.setStatus(ResponseCode.FAILURE_CODE);
 				resp.setMessage(e.getMessage());
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<serverResp>(resp, HttpStatus.ACCEPTED);
 			}
-
 		} else {
-			resp.setStatus(WebConstants.FAILURE_CODE);
-			resp.setMessage(WebConstants.FAILURE_MESSAGE);
-			return new ResponseEntity<serverResp>(resp, HttpStatus.NOT_ACCEPTABLE);
+			resp.setStatus(ResponseCode.FAILURE_CODE);
+			resp.setMessage(ResponseCode.FAILURE_MESSAGE);
 		}
+		return new ResponseEntity<serverResp>(resp, HttpStatus.ACCEPTED);
 	}
 
 	@GetMapping("/delProduct")
 	public ResponseEntity<prodResp> delProduct(@RequestHeader(name = WebConstants.USER_AUTH_TOKEN) String AUTH_TOKEN,
 			@RequestParam(name = WebConstants.PROD_ID) String productid) throws IOException {
 		prodResp resp = new prodResp();
-		if (jwtutil.checkToken(AUTH_TOKEN) != null) {
+		if (Validator.isStringEmpty(productid)) {
+			resp.setStatus(ResponseCode.BAD_REQUEST_CODE);
+			resp.setMessage(ResponseCode.BAD_REQUEST_MESSAGE);
+		} else if (!Validator.isStringEmpty(AUTH_TOKEN) && jwtutil.checkToken(AUTH_TOKEN) != null) {
 			try {
 				prodRepo.deleteByProductid(Integer.parseInt(productid));
-				resp.setStatus(WebConstants.SUCCESS_CODE);
-				resp.setMessage(WebConstants.DEL_SUCCESS_MESSAGE);
+				resp.setStatus(ResponseCode.SUCCESS_CODE);
+				resp.setMessage(ResponseCode.DEL_SUCCESS_MESSAGE);
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
 				resp.setOblist(prodRepo.findAll());
-				return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 			} catch (Exception e) {
-				resp.setStatus(WebConstants.DEL_FAILURE_CODE);
+				resp.setStatus(ResponseCode.FAILURE_CODE);
 				resp.setMessage(e.toString());
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 			}
 		} else {
-			resp.setStatus(WebConstants.FAILURE_CODE);
-			resp.setMessage(WebConstants.FAILURE_MESSAGE);
-			return new ResponseEntity<prodResp>(resp, HttpStatus.NOT_ACCEPTABLE);
+			resp.setStatus(ResponseCode.FAILURE_CODE);
+			resp.setMessage(ResponseCode.FAILURE_MESSAGE);
 		}
+		return new ResponseEntity<prodResp>(resp, HttpStatus.ACCEPTED);
 	}
 
 	@GetMapping("/viewOrders")
@@ -200,37 +213,32 @@ public class AdminController {
 			throws IOException {
 
 		viewOrdResp resp = new viewOrdResp();
-		if (jwtutil.checkToken(AUTH_TOKEN) != null) {
+		if (!Validator.isStringEmpty(AUTH_TOKEN) && jwtutil.checkToken(AUTH_TOKEN) != null) {
 			try {
-				resp.setStatus(WebConstants.SUCCESS_CODE);
-				resp.setMessage(WebConstants.VIEW_SUCCESS_MESSAGE);
+				resp.setStatus(ResponseCode.SUCCESS_CODE);
+				resp.setMessage(ResponseCode.VIEW_SUCCESS_MESSAGE);
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
 				List<order> orderList = new ArrayList<>();
-				// cartRepo;
-
-				order ord;
 				List<PlaceOrder> poList = ordRepo.findAll();
-				for (PlaceOrder p : poList) {
-					ord = new order();
-					ord.setOrderBy(p.getEmail());
-					ord.setOrderId(p.getOrderId());
-					ord.setOrderStatus(p.getOrderStatus());
-					ord.setProducts(cartRepo.findAllByOrderId(p.getOrderId()));
+				poList.forEach((po) -> {
+					order ord = new order();
+					ord.setOrderBy(po.getEmail());
+					ord.setOrderId(po.getOrderId());
+					ord.setOrderStatus(po.getOrderStatus());
+					ord.setProducts(cartRepo.findAllByOrderId(po.getOrderId()));
 					orderList.add(ord);
-				}
+				});
 				resp.setOrderlist(orderList);
-				return new ResponseEntity<viewOrdResp>(resp, HttpStatus.ACCEPTED);
 			} catch (Exception e) {
-				resp.setStatus(WebConstants.VIEW_FAILURE_CODE);
+				resp.setStatus(ResponseCode.FAILURE_CODE);
 				resp.setMessage(e.getMessage());
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<viewOrdResp>(resp, HttpStatus.ACCEPTED);
 			}
 		} else {
-			resp.setStatus(WebConstants.FAILURE_CODE);
-			resp.setMessage(WebConstants.FAILURE_MESSAGE);
-			return new ResponseEntity<viewOrdResp>(HttpStatus.NOT_ACCEPTABLE);
+			resp.setStatus(ResponseCode.FAILURE_CODE);
+			resp.setMessage(ResponseCode.FAILURE_MESSAGE);
 		}
+		return new ResponseEntity<viewOrdResp>(resp, HttpStatus.ACCEPTED);
 	}
 
 	@PostMapping("/updateOrder")
@@ -240,25 +248,26 @@ public class AdminController {
 			@RequestParam(name = WebConstants.ORD_STATUS) String orderStatus) throws IOException {
 
 		serverResp resp = new serverResp();
-		if (jwtutil.checkToken(AUTH_TOKEN) != null) {
+		if (Validator.isStringEmpty(orderId) || Validator.isStringEmpty(orderStatus)) {
+			resp.setStatus(ResponseCode.BAD_REQUEST_CODE);
+			resp.setMessage(ResponseCode.BAD_REQUEST_MESSAGE);
+		} else if (!Validator.isStringEmpty(AUTH_TOKEN) && jwtutil.checkToken(AUTH_TOKEN) != null) {
 			try {
 				PlaceOrder pc = ordRepo.findByOrderId(Integer.parseInt(orderId));
 				pc.setOrderStatus(orderStatus);
 				ordRepo.save(pc);
-				resp.setStatus(WebConstants.SUCCESS_CODE);
-				resp.setMessage(WebConstants.UPD_ORD_SUCCESS_MESSAGE);
+				resp.setStatus(ResponseCode.SUCCESS_CODE);
+				resp.setMessage(ResponseCode.UPD_ORD_SUCCESS_MESSAGE);
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<serverResp>(resp, HttpStatus.ACCEPTED);
 			} catch (Exception e) {
-				resp.setStatus(WebConstants.UPD_ORD_FAILURE_CODE);
+				resp.setStatus(ResponseCode.FAILURE_CODE);
 				resp.setMessage(e.toString());
 				resp.setAUTH_TOKEN(AUTH_TOKEN);
-				return new ResponseEntity<serverResp>(resp, HttpStatus.ACCEPTED);
 			}
 		} else {
-			resp.setStatus(WebConstants.FAILURE_CODE);
-			resp.setMessage(WebConstants.FAILURE_MESSAGE);
-			return new ResponseEntity<serverResp>(HttpStatus.NOT_ACCEPTABLE);
+			resp.setStatus(ResponseCode.FAILURE_CODE);
+			resp.setMessage(ResponseCode.FAILURE_MESSAGE);
 		}
+		return new ResponseEntity<serverResp>(resp, HttpStatus.ACCEPTED);
 	}
 }
